# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.LEIRecord:
    properties:
      changed_fields:
        description: Last change details
        type: string
      child_lei:
        type: string
      created_at:
        description: Standard fields
        type: string
      created_by:
        type: string
      entity_category:
        type: string
      entity_legal_form:
        type: string
      entity_legal_form_name:
        type: string
      entity_status:
        type: string
      entity_sub_category:
        type: string
      hq_address_city:
        type: string
      hq_address_country:
        description: ISO 3166-1 alpha-2
        type: string
      hq_address_line_1:
        description: Headquarters address
        type: string
      hq_address_line_2:
        type: string
      hq_address_line_3:
        type: string
      hq_address_line_4:
        type: string
      hq_address_postal_code:
        type: string
      hq_address_region:
        type: string
      id:
        type: string
      initial_registration_date:
        description: Dates
        type: string
      is_provisional:
        description: |-
          Provisional LEI fields (migration 000062)
          IsProvisional is TRUE for Axiom-issued provisional LEI records (AXIO prefix).
          FALSE for all GLEIF-ingested records. Defaults to FALSE.
        type: boolean
      last_update_date:
        type: string
      legal_address_city:
        type: string
      legal_address_country:
        description: ISO 3166-1 alpha-2
        type: string
      legal_address_line_1:
        description: Legal address
        type: string
      legal_address_line_2:
        type: string
      legal_address_line_3:
        type: string
      legal_address_line_4:
        type: string
      legal_address_postal_code:
        type: string
      legal_address_region:
        type: string
      legal_jurisdiction:
        type: string
      legal_name:
        description: Entity information
        type: string
      lei:
        description: Legal Entity Identifier (unique)
        type: string
      managing_lou:
        description: Associated entities
        type: string
      managing_lou_legal_name:
        type: string
      next_renewal_date:
        type: string
      other_names:
        description: Array of alternative names
        type: string
      parent_lei:
        description: |-
          Relationship data hydrated from lei_relationship_records for provisional LEIs.
          Not stored in lei_records; populated by the provisional LEI service layer only.
        type: string
      provisioning_source:
        type: string
      registration_authority:
        description: Registration
        type: string
      registration_authority_comments:
        type: string
      registration_authority_id:
        type: string
      registration_authority_international_name:
        type: string
      registration_authority_name:
        description: GLEIF reference resolved names (computed via JOIN, not stored
          in lei_records)
        type: string
      registration_authority_website:
        type: string
      registration_number:
        type: string
      registration_status:
        type: string
      source_file:
        $ref: '#/definitions/domain.SourceFile'
      source_file_id:
        description: Audit and provenance
        type: string
      successor_lei:
        type: string
      successor_lei_legal_name:
        type: string
      transliterated_legal_name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      validation_authority:
        type: string
      validation_sources:
        description: Validation
        type: string
    required:
    - lei
    type: object
...
  domain.SourceFile:
    properties:
      created_at:
        type: string
      downloaded_at:
        type: string
      failed_records:
        type: integer
      failure_category:
        description: SCHEMA_ERROR, NETWORK_ERROR, FILE_CORRUPTION, UNKNOWN
        type: string
      file_hash:
        description: SHA-256 hash
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        description: FULL, DELTA
        type: string
      file_url:
        type: string
      id:
        type: string
      job_label:
        type: string
      job_type:
        type: string
      last_processed_lei:
        description: For resumption
        type: string
      max_retries:
        type: integer
      processed_records:
        type: integer
      processing_completed_at:
        type: string
      processing_error:
        type: string
      processing_started_at:
        type: string
      processing_status:
        description: Processing status
        type: string
      publication_date:
        type: string
      retry_count:
        description: Retry tracking
        type: integer
      total_records:
        type: integer
      updated_at:
        type: string
    type: object
...
  handler.succeedProvisionalLEIRequest:
    properties:
      official_lei:
        type: string
    required:
    - official_lei
    type: object
...
  service.CreateProvisionalLEIRequest:
    properties:
      child_lei:
        type: string
      legal_address_city:
        type: string
      legal_address_country:
        type: string
      legal_jurisdiction:
        type: string
      legal_name:
        type: string
      notes:
        type: string
      parent_lei:
        type: string
      provisioning_source:
        description: e.g. "onboarding", "counterparty"
        type: string
    required:
    - legal_name
    type: object
...
  service.UpdateProvisionalLEIRequest:
    properties:
      child_lei:
        type: string
      entity_status:
        type: string
      legal_address_city:
        type: string
      legal_address_country:
        type: string
      legal_jurisdiction:
        type: string
      legal_name:
        type: string
      parent_lei:
        type: string
      provisioning_source:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@axiom.example.com
    name: API Support
  description: Financial Services Static Data Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Axiom API
  version: "1.0"
paths:
...
  /api/v1/lei/provisional:
    get:
      description: Returns provisional LEI records with pagination.
      parameters:
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List provisional LEI records
      tags:
      - Provisional LEI
    post:
      consumes:
      - application/json
      description: Issues a new provisional LEI record.
      parameters:
      - description: Create provisional LEI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateProvisionalLEIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.LEIRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create provisional LEI record
      tags:
      - Provisional LEI
  /api/v1/lei/provisional/{lei}:
    get:
      description: Returns a single provisional LEI record by LEI code.
      parameters:
      - description: Provisional LEI code
        in: path
        name: lei
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LEIRecord'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get provisional LEI record
      tags:
      - Provisional LEI
    put:
      consumes:
      - application/json
      description: Updates mutable fields of an existing provisional LEI record.
      parameters:
      - description: Provisional LEI code
        in: path
        name: lei
        required: true
        type: string
      - description: Update provisional LEI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateProvisionalLEIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LEIRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update provisional LEI record
      tags:
      - Provisional LEI
  /api/v1/lei/provisional/{lei}/succeed:
    post:
      consumes:
      - application/json
      description: Marks a provisional LEI as succeeded by a successor LEI.
      parameters:
      - description: Provisional LEI code
        in: path
        name: lei
        required: true
        type: string
      - description: Succeed provisional LEI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.succeedProvisionalLEIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Succeed provisional LEI
      tags:
      - Provisional LEI
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"